Fixed a weird bug where ActionListeners were never unreged
Required me making a new function, unRegister(), in Menu that is
implemented in ActionMenu. It took the place of two functions in
ActionMenu that didn't always catch the transition of menus. I also made
"exit" get handled by the Menu.getNextPrompt(Context) so that I could be
sure that listeners were cancelled when HUD was closed.