Fixed some bugs with manageFacility not removing facils
Also, I made a dramatic change to how facilities are stored. For some
reason governments had a list of all the facilities stored locally. But,
the facilities were also stored in the regiondata. I decided to do away
with this redundancy and only have them exist in the regiondata.
Governments can get to them in the same way that they get all the houses
or towns.  In addition, I added a function to the Multi map that allows
for easy removal of single values.