Implementing NPC Buying. Fixed so many bugs.
I added an output file that translates the entire item tree into a yml
file. This will be substantially easier to navigate and potentially
easier to read, once I get code folding to work with yml files. NPC can
buy a good amount of items now, I've gotten rid of a a lot of the debug
statements.