Fixed a division by zero in TabSelectOption Menu
When a player tried to search another term after one of their search
terms returned zero results, a division by zero occured. This is because
the list of search terms was empty, and the position of the tab pointer
was determined by dividing the list index by the length of the list.