Finally sign shops are claimable.
Fixed a few bugs. There seems to be an issue still with the plugin
randomly not being able to find a gov. I thought I had fixed it by
changning multipmaps to have lists instead of HashSets, but it still
seems to be doing it. I'll keep my eyes open.